refactor: function in .h move to .cc & change #define to constexpr
