+++ fw/Core/Hitcon/App/tama_src/simu_setting.h
@@ -3,7 +3,7 @@
 
 typedef unsigned char uint8_t;
 
-#include <unistd.h>
-
 #include <iostream>
 #endif
+
+#include <unistd.h>
